requirements.txt
chromadb==1.4.0
groq==1.0.0
python-dotenv==1.2.1
